requirements.txt
fastapi
uvicorn
langchain
langgraph
faiss-cpu
groq
python-dotenv